+++ Tests/Serialization/TypeSerializers/TransactionSerializerTests.swift
@@ -17,12 +17,12 @@ class TransactionSerializerTests: XCTestCase {
             outputs: [
                 CellOutput(
                     capacity: 100000000000,
-                    lock: Script(args: ["0xe2193df51d78411601796b35b17b4f8f2cd85bd0"], codeHash: "0x9e3b3557f11b2b3532ce352bfe8017e9fd11d154c4c7f9b7aaaa1e621b539a08"),
+                    lock: Script(args: "0xe2193df51d78411601796b35b17b4f8f2cd85bd0", codeHash: "0x9e3b3557f11b2b3532ce352bfe8017e9fd11d154c4c7f9b7aaaa1e621b539a08"),
                     type: nil
                 ),
                 CellOutput(
                     capacity: 4900000000000,
-                    lock: Script(args: ["0x36c329ed630d6ce750712a477543672adab57f4c"], codeHash: "0x9e3b3557f11b2b3532ce352bfe8017e9fd11d154c4c7f9b7aaaa1e621b539a08"),
+                    lock: Script(args: "0x36c329ed630d6ce750712a477543672adab57f4c", codeHash: "0x9e3b3557f11b2b3532ce352bfe8017e9fd11d154c4c7f9b7aaaa1e621b539a08"),
                     type: nil
                 )
             ],
@@ -31,10 +31,10 @@ class TransactionSerializerTests: XCTestCase {
                 "0x"
             ],
             witnesses: [
-                Witness(data: [])
+                "0x"
             ]
         )
-        XCTAssertEqual(tx.computeHash(), "0xc932addf21edeac2fdd4d677f3984688ee5514c87412087c6bb533bf96e4e624")
+        XCTAssertEqual(tx.computeHash(), "0x4c02905db773301f73bbc6cd5a400c928caf410bbb13136f6f48bec0a79c22e4")
         // let hash = try? APIClient(url: APIClient.defaultLocalURL).computeTransactionHash(transaction: tx)
         // XCTAssertEqual(tx.computeHash(), hash)
     }
@@ -44,11 +44,29 @@ class TransactionSerializerTests: XCTestCase {
             cellDeps: [CellDep(outPoint: OutPoint(txHash: "0x29f94532fb6c7a17f13bcde5adb6e2921776ee6f357adf645e5393bd13442141", index: 0), depType: .code)],
             headerDeps: ["0x8033e126475d197f2366bbc2f30b907d15af85c9d9533253c6f0787dcbbb509e"],
             inputs: [CellInput(previousOutput: OutPoint(txHash: "0x5ba156200c6310bf140fbbd3bfe7e8f03d4d5f82b612c1a8ec2501826eaabc17", index: 0), since: 0)],
-            outputs: [CellOutput(capacity: 100000000000, lock: Script(args: [], codeHash: "0x28e83a1277d48add8e72fadaa9248559e1b632bab2bd60b27955ebc4c03800a5", hashType: .data))],
+            outputs: [CellOutput(capacity: 100000000000, lock: Script(args: "0x", codeHash: "0x28e83a1277d48add8e72fadaa9248559e1b632bab2bd60b27955ebc4c03800a5", hashType: .data))],
             outputsData: ["0x"]
         )
-        XCTAssertEqual(tx.computeHash(), "0xba86cc2cb21832bf4a84c032eb6e8dc422385cc8f8efb84eb0bc5fe0b0b9aece")
+        XCTAssertEqual(tx.computeHash(), "0x6f6b16079884d8127490aac5e0f87274e81f15ea1fd6c9672a5b0326bd8ce76d")
         // let hash = try? APIClient(url: APIClient.defaultLocalURL).computeTransactionHash(transaction: tx)
         // XCTAssertEqual(tx.computeHash(), hash)
+    }
+
+    func testTransactionWithWitnesses() {
+        let tx = Transaction(
+            cellDeps: [CellDep(outPoint: OutPoint(txHash: "0x29f94532fb6c7a17f13bcde5adb6e2921776ee6f357adf645e5393bd13442141", index: 0), depType: .code)],
+            headerDeps: ["0x8033e126475d197f2366bbc2f30b907d15af85c9d9533253c6f0787dcbbb509e"],
+            inputs: [
+                CellInput(previousOutput: OutPoint(txHash: "0x5ba156200c6310bf140fbbd3bfe7e8f03d4d5f82b612c1a8ec2501826eaabc17", index: 0), since: 0),
+                CellInput(previousOutput: OutPoint(txHash: "0x5ba156200c6310bf140fbbd3bfe7e8f03d4d5f82b612c1a8ec2501826eaabc17", index: 0), since: 1)
+            ],
+            outputs: [],
+            outputsData: ["0x"],
+            witnesses: [
+                "0x4107bd23eedb9f2a2a749108f6bb9720d745d50f044cc4814bafe189a01fe6fb",
+                "0x4107bd23eedb9f2a2a749108f6bb9720d745d50f044cc4814bafe189a01fe6fb"
+            ]
+        )
+        XCTAssertEqual(tx.computeHash(), "0xe37f68d4d81e0f790ca98c5abbd763764f826170867a0b96137f68df908e5641")
     }
 }